rename tools module
